check if package exists in database before try to delete file

+++ frontend/views/package.php
@@ -10,9 +10,11 @@ if(!empty($_POST['move_in_group']) && !empty($_POST['move_from_pos']) && !empty(
 if(!empty($_POST['remove_id']) && is_array($_POST['remove_id'])) {
 	foreach($_POST['remove_id'] as $id) {
 		$package = $db->getPackage($id);
-		$path = PACKAGE_PATH.'/'.intval($package->id).'.zip';
-		unlink($path);
-		$db->removePackage($package->id);
+		if(!empty($package)) {
+			$path = PACKAGE_PATH.'/'.intval($package->id).'.zip';
+			unlink($path);
+			$db->removePackage($package->id);
+		}
 	}
 	die();
 }
